a number of small additions
