+++ package.json
@@ -28,6 +28,6 @@
     "packages/*"
   ],
   "gitHooks": {
-    "pre-commit": "lerna run --concurrency 1 --stream precommit --since HEAD"
+    "pre-commit": "lerna run precommit --concurrency=1 --stream --since HEAD"
   }
 }
