+++ packages/docs/plugins/color-helpers.ts
@@ -1,4 +1,3 @@
-
 import Vue from 'vue'
 import ColorHelpersPlugin from '../../ui/src/components/vuestic-utilities/color-helpers-plugin'
