+++ packages/docs/plugins/color-theme.ts
@@ -1,0 +1,4 @@
+import { setTheme } from '../../ui/src/services/ColorThemePlugin'
+import { COLOR_THEMES, ThemeName } from '../theme-config'
+
+setTheme(COLOR_THEMES[ThemeName.DEFAULT])
