+++ packages/docs/plugins/external-vuetable.ts
@@ -1,5 +1,4 @@
 // @ts-nocheck
-
 import Vue from 'vue'
 import Vuetable from 'vuetable-2/src/components/Vuetable.vue'
 Vue.use(Vuetable)
