+++ packages/ui/src/components/vuestic-components/va-icon/VaIcon.vue
@@ -13,11 +13,11 @@
 </template>
 
 <script>
-import { ColorThemeMixin } from '../../../services/ColorThemePlugin'
 import { SizeMixin } from '../../../mixins/SizeMixin'
 import { makeContextablePropsMixin } from '../../context-test/context-provide/ContextPlugin'
 import vaIconMixin from './vaIconMixin'
 import { warn } from '../../../services/utils'
+import { ColorThemeMixin } from '../../vuestic-mixins/ColorMixin'
 
 const IconContextMixin = makeContextablePropsMixin({
   name: {
