+++ packages/ui/src/components/vuestic-components/va-input/VaInput.vue
@@ -96,10 +96,10 @@ import VaInputWrapper from '../va-input/VaInputWrapper'
 import VaIcon from '../va-icon/VaIcon'
 import { getHoverColor } from '../../../services/color-functions'
 import calculateNodeHeight from './calculateNodeHeight'
-import { ColorThemeMixin } from '../../../services/ColorThemePlugin'
 import { makeContextablePropsMixin } from '../../context-test/context-provide/ContextPlugin'
 import { FormComponentMixin } from '../../vuestic-mixins/FormComponent/FormComponentMixin'
 import { warn } from '../../../services/utils'
+import { ColorThemeMixin } from '../../vuestic-mixins/ColorMixin'
 
 const InputContextMixin = makeContextablePropsMixin({
   color: {
