+++ packages/ui/src/components/vuestic-components/va-list/VaListLabel.vue
@@ -8,7 +8,7 @@
 </template>
 
 <script>
-import { ColorThemeMixin } from '../../../services/ColorThemePlugin'
+import { ColorThemeMixin } from '../../vuestic-mixins/ColorMixin'
 
 export default {
   name: 'VaListLabel',
