+++ packages/ui/src/components/vuestic-components/va-popover/VaPopover.vue
@@ -46,7 +46,6 @@ import {
   getHoverColor,
   getBoxShadowColor,
 } from '../../../services/color-functions'
-import { getColor } from '../../../services/ColorThemePlugin'
 
 export default {
   name: 'VaPopover',
