+++ packages/ui/src/components/vuestic-components/va-progress-bar/progress-types/VaProgressBar.vue
@@ -47,9 +47,9 @@
 <script>
 import { progressMixin } from './progressMixin'
 import { normalizeValue } from '../../../../services/utils'
-import { ColorThemeMixin } from '../../../../services/ColorThemePlugin'
 import { makeContextablePropsMixin } from '../../../context-test/context-provide/ContextPlugin'
 import { SizeMixin } from '../../../../mixins/SizeMixin'
+import { ColorThemeMixin } from '../../../vuestic-mixins/ColorMixin'
 
 export default {
   name: 'VaProgressBar',
