+++ packages/ui/src/components/vuestic-components/va-radio/VaRadio.vue
@@ -37,8 +37,8 @@
 </template>
 
 <script>
-import { ColorThemeMixin } from '../../../services/ColorThemePlugin'
 import { makeContextablePropsMixin } from '../../context-test/context-provide/ContextPlugin'
+import { ColorThemeMixin } from '../../vuestic-mixins/ColorMixin'
 
 const RadioContextMixin = makeContextablePropsMixin({
   value: {
