+++ packages/ui/src/components/vuestic-utilities/color-helpers-plugin.ts
@@ -2,6 +2,12 @@ import Vue from 'vue'
 import { themesRef } from '../../services/ColorThemePlugin'
 import { addOrUpdateStyleElement } from '../../services/dom-functions'
 
+declare module 'vue/types/vue' {
+  interface Vue {
+    $themes: Record<string, string> | undefined;
+  }
+}
+
 const vmInstance = new Vue()
 
 const createThemeColorStyles = (themes: Record<string, string>): string => {
@@ -17,8 +23,19 @@ const createThemeColorStyles = (themes: Record<string, string>): string => {
   return result
 }
 
+// This plugin is considered a separate entity from components.
+// Components should not be aware it exists at all. Neither it should be used in vue-book.
+// Appropriate use cases are user app and documentation.
 const ColorHelpersPlugin = {
   install () {
+    // NOTE Nuxt went broken without this check. Seemingly loaded plugin twice.
+    if (!Vue.prototype.$themes) {
+      // @ts-ignore
+      Object.defineProperty(Vue.prototype, '$themes', {
+        get: () => themesRef.value,
+      })
+    }
+
     // @ts-ignore
     vmInstance.$watch(() => themesRef.value, {
       deep: true,
