+++ packages/ui/src/services/ColorThemePlugin-docs.md
@@ -5,7 +5,9 @@ This plugin provides app-wide colors.
 You can define your own colors or change existing ones by providing them via plugin options:
 
 ```js
-Vue.use(ColorThemePlugin, {themes: {primary: '#f06595', blurple: '#7289DA'}})
+import { setTheme } from 'vuestic-ui'
+
+setTheme({primary: '#f06595', blurple: '#7289DA'})
 ```
 
 And you can use them in component code or template like this:
