+++ packages/ui/src/services/ColorThemePlugin.ts
@@ -1,12 +1,4 @@
 import Vue from 'vue'
-import { makeContextablePropsMixin } from '../components/context-test/context-provide/ContextPlugin'
-import { Component } from 'vue-property-decorator'
-
-declare module 'vue/types/vue' {
-  interface Vue {
-    $themes: Record<string, string> | undefined;
-  }
-}
 
 // Most default color - fallback when nothing else is found.
 const DEFAULT_COLOR = '#000000'
@@ -25,16 +17,12 @@ export const themesRef: {value: Record<string, string>} = Vue.observable({
   },
 })
 
+// @deprecated
 export const ColorThemePlugin = {
   install (Vue: Vue, options?: { themes?: Record<string, string> }) {
     if (options && options.themes) {
       themesRef.value = { ...themesRef.value, ...options.themes }
     }
-    // @ts-ignore
-    // This plugin is not intended.
-    Object.defineProperty(Vue.prototype, '$themes', {
-      get: () => themesRef.value,
-    })
   },
 }
 
@@ -65,27 +53,4 @@ export const getColor = ($vm: Vue, prop: string, defaultColor: string = DEFAULT_
 
 export const setTheme = (theme: Record<string, string>): void => {
   Object.assign(themesRef.value, theme)
-}
-
-/**
- * This mixin does not cover all needed color functionality for majority of components.
- * Its main purpose is raw code reuse.
- */
-@Component({
-  mixins: [
-    makeContextablePropsMixin({
-      color: {
-        type: String,
-      },
-    }),
-  ] as any,
-})
-export class ColorThemeMixin extends Vue {
-  get colorComputed () {
-    return getColor(this, (this as any).c_color)
-  }
-
-  computeColor (prop: string, defaultColor?: string) {
-    return getColor(this, prop, defaultColor)
-  }
 }
