--- packages/ui/src/services/color-themes.ts
@@ -1,33 +1,0 @@
-export const THEME_NAMES = {
-  DEFAULT: 'DEFAULT',
-  CORPORATE: 'CORPORATE',
-}
-// Ready theme presets.
-export const COLOR_THEMES = [
-  {
-    name: THEME_NAMES.DEFAULT,
-    themes: {
-      primary: '#40e583',
-      secondary: '#002c85',
-      success: '#40e583',
-      info: '#2c82e0',
-      danger: '#e34b4a',
-      warning: '#ffc200',
-      gray: '#babfc2',
-      dark: '#34495e',
-    },
-  },
-  {
-    name: THEME_NAMES.CORPORATE,
-    themes: {
-      primary: '#6c7fee',
-      secondary: '#6e7ff1',
-      success: '#8ddc88',
-      info: '#71baff',
-      danger: '#f8706d',
-      warning: '#ffd652',
-      gray: '#8396a5',
-      dark: '#34495e',
-    },
-  },
-]
