+++ packages/ui/src/vue-book/book-main.ts
@@ -5,7 +5,7 @@ import BookApp from './BookApp.vue'
 import VueClipboard from 'vue-clipboard2'
 import Router from 'vue-router'
 import { VueBookComponents, createRoute } from 'vue-book'
-import { ColorThemePlugin } from '../services/ColorThemePlugin'
+import { setTheme } from '../services/ColorThemePlugin'
 import { getContext } from '../components/context-test/context-provide/context'
 import { ContextPlugin } from '../components/context-test/context-provide/ContextPlugin'
 import { BusPlugin } from 'vue-epic-bus'
@@ -22,7 +22,7 @@ installPlatform()
 Vue.use(Router)
 Vue.use(VueBookComponents)
 if (!process.env.VUE_APP_DEMO_NO_THEME_PLUGIN) {
-  Vue.use(ColorThemePlugin)
+  setTheme({ primary: '#e71919' })
 }
 Vue.use(DropdownPopperPlugin)
 Vue.use(ContextPlugin, getContext())
